refactor(ui): tabbed Account page + extracted section components + best-times on Dashboard
Behavior-preserving restructure of the Account page. No API endpoints,
request/response shapes, React Query keys, or business logic changed.

- Add reusable <SettingsCard> wrapper and moved <Toggle> to its own file.
- Split the ~1500-line Account.tsx into per-section child components under
  pages/account/, each owning its own state/query/mutation.
- Account.tsx becomes a thin composition: checklist banner + 4-tab layout
  (Account & Billing / LinkedIn Connection / Content & Profile /
  Engagement & Automation), with the active tab persisted in ?tab=.
  Existing ?li_connected= / ?li_error= OAuth handling preserved.
- Move the read-only "Your Best Times to Post" panel (post-stats query)
  from Account to the Dashboard.
