+++ src/cqc_lem/ui/src/pages/Account.tsx
@@ -1,205 +1,52 @@
-import { useState, useEffect } from 'react'
+import { useEffect, useState } from 'react'
 import { useSearchParams } from 'react-router-dom'
-import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query'
-import api from '../api/client'
-import { useAuth } from '../contexts/AuthContext'
+import { useQueryClient } from '@tanstack/react-query'
 import { useAccountReadiness } from '../hooks/useAccountReadiness'
-import LinkedInSessionCard from '../components/LinkedInSessionCard'
+import SubscriptionCard from './account/SubscriptionCard'
+import VideoCreditsCard from './account/VideoCreditsCard'
+import TimezoneCard from './account/TimezoneCard'
+import LoginLocationCard from './account/LoginLocationCard'
+import InactivityCard from './account/InactivityCard'
+import LinkedInLoginCard from './account/LinkedInLoginCard'
+import CompanyPageCard from './account/CompanyPageCard'
+import ContentProfileCard from './account/ContentProfileCard'
+import NewsletterCard from './account/NewsletterCard'
+import EngagementPreferencesCard from './account/EngagementPreferencesCard'
+import EngagementTargetingCard from './account/EngagementTargetingCard'
+import GroupsCard from './account/GroupsCard'
+import DmTemplatesCard from './account/DmTemplatesCard'
+import LeadMagnetCard from './account/LeadMagnetCard'
 
-const LI_AUTH_URL = '/api/auth/linkedin/'
+type TabKey = 'account' | 'linkedin' | 'content' | 'automation'
 
-const TIER_LABELS: Record<string, string> = {
-  free_trial: 'Free Trial',
-  starter: 'Starter',
-  professional: 'Professional',
-  enterprise: 'Enterprise',
-}
-
-const TIER_COLORS: Record<string, string> = {
-  free_trial: 'bg-gray-100 text-gray-700',
-  starter: 'bg-blue-100 text-blue-800',
-  professional: 'bg-purple-100 text-purple-800',
-  enterprise: 'bg-yellow-100 text-yellow-800',
-}
-
-const VIDEO_PACKAGES = [
-  { key: 'small', price: '$12', label: '5 credits' },
-  { key: 'medium', price: '$30', label: '15 credits' },
-  { key: 'large', price: '$70', label: '40 credits' },
-  { key: 'max', price: '$150', label: '100 credits' },
+const TABS: { key: TabKey; label: string }[] = [
+  { key: 'account', label: 'Account & Billing' },
+  { key: 'linkedin', label: 'LinkedIn Connection' },
+  { key: 'content', label: 'Content & Profile' },
+  { key: 'automation', label: 'Engagement & Automation' },
 ]
 
-const COMMON_TIMEZONES = [
-  'America/New_York',
-  'America/Chicago',
-  'America/Denver',
-  'America/Los_Angeles',
-  'America/Phoenix',
-  'America/Anchorage',
-  'Pacific/Honolulu',
-  'America/Toronto',
-  'America/Vancouver',
-  'America/Sao_Paulo',
-  'America/Mexico_City',
-  'Europe/London',
-  'Europe/Paris',
-  'Europe/Berlin',
-  'Europe/Amsterdam',
-  'Europe/Rome',
-  'Europe/Madrid',
-  'Europe/Moscow',
-  'Asia/Dubai',
-  'Asia/Kolkata',
-  'Asia/Singapore',
-  'Asia/Tokyo',
-  'Asia/Shanghai',
-  'Australia/Sydney',
-  'Pacific/Auckland',
-  'UTC',
-]
-
-const INACTIVATE_OPTIONS = [
-  { label: '30 days', value: 30 },
-  { label: '60 days', value: 60 },
-  { label: '90 days (default)', value: 90 },
-  { label: '120 days', value: 120 },
-  { label: '365 days', value: 365 },
-  { label: 'Never', value: null },
-]
-
-function SectionHeading({ title }: { title: string }) {
-  return (
-    <h2 className="text-xs font-bold uppercase tracking-wide text-gray-400 pt-3 -mb-2">{title}</h2>
-  )
-}
-
-function daysUntil(iso: string | null | undefined): number | null {
-  if (!iso) return null
-  const diff = new Date(iso).getTime() - Date.now()
-  return Math.max(0, Math.floor(diff / (1000 * 60 * 60 * 24)))
-}
-
-const DM_EVENTS: { key: string; label: string }[] = [
-  { key: 'connection_accepted', label: 'Connection accepted' },
-  { key: 'recommendation_received', label: 'Recommendation received' },
-  { key: 'collaboration', label: 'After a collaboration' },
-  { key: 'profile_viewer', label: 'Profile viewer outreach' },
-]
-
-const csv = (arr: string[] | undefined | null) => (arr && arr.length ? arr.join(', ') : '')
-const parseCsv = (s: string) => s.split(',').map((x) => x.trim()).filter(Boolean)
-
-type EngPrefs = {
-  tone: string | null
-  comment_length: string
-  comment_style: string | null
-  use_emojis: boolean
-  use_hashtags: boolean
-  include_topics: string[]
-  exclude_topics: string[]
-  include_keywords: string[]
-  exclude_keywords: string[]
-  include_authors: string[]
-  exclude_authors: string[]
-  min_reactions: number | null
-  max_post_age_hours: number | null
-  reply_to_own_comments: boolean
-  max_comments_per_day: number
-  max_dms_per_day: number
-}
-
-type DmTemplate = {
-  event_type: string
-  step: number
-  delay_hours: number
-  template_text: string
-  is_active: boolean
-}
-
-type NewsletterSettings = {
-  enabled: boolean
-  title: string | null
-  topic: string | null
-  cadence: string
-  align_with_blog: boolean
-}
-
-type UserGroup = {
-  group_id: string
-  group_name: string | null
-  enabled: boolean
-}
-
-type LeadMagnet = {
-  enabled: boolean
-  keyword: string | null
-  message: string | null
-}
-
-function Toggle({ on, onClick }: { on: boolean; onClick: () => void }) {
-  return (
-    <button
-      type="button"
-      onClick={onClick}
-      className={`relative inline-flex h-6 w-11 flex-shrink-0 cursor-pointer rounded-full border-2 border-transparent transition-colors duration-200 focus:outline-none focus:ring-2 focus:ring-blue-500 focus:ring-offset-2 ${on ? 'bg-blue-600' : 'bg-gray-200'}`}
-      role="switch"
-      aria-checked={on}
-    >
-      <span className={`pointer-events-none inline-block h-5 w-5 transform rounded-full bg-white shadow ring-0 transition duration-200 ease-in-out ${on ? 'translate-x-5' : 'translate-x-0'}`} />
-    </button>
-  )
-}
+const VALID_TABS = TABS.map((t) => t.key) as string[]
 
 export default function Account() {
-  const { user, sessionToken } = useAuth()
-  const email = user?.email ?? ''
   const queryClient = useQueryClient()
   const { data: readiness } = useAccountReadiness()
-  const sessionOk = readiness?.items.find((i) => i.key === 'linkedin_session')?.ok ?? false
-
   const [searchParams, setSearchParams] = useSearchParams()
-  const [blogUrl, setBlogUrl] = useState(localStorage.getItem('lem_blog_url') || '')
-  const [sitemapUrl, setSitemapUrl] = useState(localStorage.getItem('lem_sitemap_url') || '')
-  const [companyPage, setCompanyPage] = useState('')
-  const [companyPageMsg, setCompanyPageMsg] = useState<{ ok: boolean; text: string } | null>(null)
-  const [liConnectedLocal, setLiConnectedLocal] = useState(
-    localStorage.getItem('lem_li_connected') === '1'
-  )
-  const [savedMsg, setSavedMsg] = useState<string | null>(null)
-  const [prefsSavedMsg, setPrefsSavedMsg] = useState<string | null>(null)
-  const [checkoutMsg, setCheckoutMsg] = useState<{ ok: boolean; text: string } | null>(null)
-  const [liPassword, setLiPassword] = useState('')
-  const [showLiPassword, setShowLiPassword] = useState(false)
-  const [liPasswordMsg, setLiPasswordMsg] = useState<{ ok: boolean; text: string } | null>(null)
+  const [oauthMsg, setOauthMsg] = useState<{ ok: boolean; text: string } | null>(null)
 
-  // Preferences local state — initialised from query once loaded
-  const [inactivateDelay, setInactivateDelay] = useState<number | null>(90)
-  const [autoSchedule, setAutoSchedule] = useState(false)
-  const [prefsInitialised, setPrefsInitialised] = useState(false)
-  const [urlsInitialised, setUrlsInitialised] = useState(false)
+  const tabParam = searchParams.get('tab')
+  const activeTab: TabKey = VALID_TABS.includes(tabParam ?? '') ? (tabParam as TabKey) : 'account'
 
-  // Timezone state
-  const [timezone, setTimezone] = useState('America/New_York')
-  const [tzInitialised, setTzInitialised] = useState(false)
-  const [tzSavedMsg, setTzSavedMsg] = useState<string | null>(null)
-  const [locationMsg, setLocationMsg] = useState<{ ok: boolean; text: string } | null>(null)
-  const [locCity, setLocCity] = useState('')
-  const [locState, setLocState] = useState('')
-  const [locCountry, setLocCountry] = useState('US')
-
-  // Engagement config (targeting + voice) and DM templates
-  const [engPrefs, setEngPrefs] = useState<EngPrefs | null>(null)
-  const [engMsg, setEngMsg] = useState<{ ok: boolean; text: string } | null>(null)
-  const [dmTemplates, setDmTemplates] = useState<DmTemplate[]>([])
-  const [dmInit, setDmInit] = useState(false)
-  const [dmMsg, setDmMsg] = useState<{ ok: boolean; text: string } | null>(null)
-  const [newsletter, setNewsletter] = useState<NewsletterSettings | null>(null)
-  const [nlMsg, setNlMsg] = useState<{ ok: boolean; text: string } | null>(null)
-  const [groups, setGroups] = useState<UserGroup[]>([])
-  const [groupsInit, setGroupsInit] = useState(false)
-  const [groupsMsg, setGroupsMsg] = useState<{ ok: boolean; text: string } | null>(null)
-  const [leadMagnet, setLeadMagnet] = useState<LeadMagnet | null>(null)
-  const [lmMsg, setLmMsg] = useState<{ ok: boolean; text: string } | null>(null)
+  const setActiveTab = (tab: TabKey) => {
+    setSearchParams(
+      (prev) => {
+        const next = new URLSearchParams(prev)
+        next.set('tab', tab)
+        return next
+      },
+      { replace: true }
+    )
+  }
 
   // Handle LinkedIn OAuth callback: ?li_connected=1 or ?li_error=... in URL
   useEffect(() => {
@@ -207,519 +54,27 @@ export default function Account() {
     const liError = searchParams.get('li_error')
     if (liConnected === '1') {
       localStorage.setItem('lem_li_connected', '1')
-      setLiConnectedLocal(true)
       // Force refetch so the fresh token is reflected immediately
       queryClient.invalidateQueries({ queryKey: ['token-status'] })
       setSearchParams({}, { replace: true })
-      setSavedMsg('LinkedIn connected successfully!')
-      setTimeout(() => setSavedMsg(null), 5000)
+      setOauthMsg({ ok: true, text: 'LinkedIn connected successfully!' })
+      setTimeout(() => setOauthMsg(null), 5000)
     } else if (liError) {
-      setSavedMsg('LinkedIn connection failed — please try again.')
+      setOauthMsg({ ok: false, text: 'LinkedIn connection failed — please try again.' })
       setSearchParams({}, { replace: true })
-      setTimeout(() => setSavedMsg(null), 8000)
+      setTimeout(() => setOauthMsg(null), 8000)
     }
-    setBlogUrl(localStorage.getItem('lem_blog_url') || '')
-    setSitemapUrl(localStorage.getItem('lem_sitemap_url') || '')
   }, [])
-
-  // LinkedIn token status
-  const { data: tokenStatusData } = useQuery({
-    queryKey: ['token-status', sessionToken],
-    queryFn: () =>
-      api
-        .get(`/user/token_status?session_token=${encodeURIComponent(sessionToken!)}`)
-        .then((r) => r.data.detail as {
-          token_expiry_date: string | null
-          is_expiring_soon: boolean
-          is_expired: boolean
-          refresh_attempted: boolean
-          refresh_succeeded: boolean
-        }),
-    enabled: !!sessionToken,
-    staleTime: 5 * 60 * 1000,
-  })
-
-  // Subscription + preferences
-  const { data: settingsData } = useQuery({
-    queryKey: ['user-settings', sessionToken],
-    queryFn: () =>
-      api
-        .get(`/user/settings?session_token=${encodeURIComponent(sessionToken!)}`)
-        .then((r) => r.data.detail as {
-          subscription: {
-            status: string | null
-            tier: string | null
-            trial_started_at: string | null
-            trial_ends_at: string | null
-            stripe_customer_id: string | null
-          } | null
-          preferences: {
-            last_login_inactivate_delay: number | null
-            auto_schedule_posts: boolean
-          } | null
-          blog_url: string | null
-          sitemap_url: string | null
-          company_linked_in_url: string | null
-        }),
-    enabled: !!sessionToken,
-    staleTime: 60 * 1000,
-  })
-
-  useEffect(() => {
-    if (settingsData?.preferences && !prefsInitialised) {
-      setInactivateDelay(settingsData.preferences.last_login_inactivate_delay)
-      setAutoSchedule(settingsData.preferences.auto_schedule_posts)
-      setPrefsInitialised(true)
-    }
-  }, [settingsData, prefsInitialised])
-
-  // Timezone query
-  const { data: tzData } = useQuery({
-    queryKey: ['user-timezone', sessionToken],
-    queryFn: () =>
-      api
-        .get(`/user/timezone?session_token=${encodeURIComponent(sessionToken!)}`)
-        .then((r) => r.data.detail as { timezone: string }),
-    enabled: !!sessionToken,
-    staleTime: 5 * 60 * 1000,
-  })
-
-  const { data: videoCreditData } = useQuery({
-    queryKey: ['video-credits', sessionToken],
-    queryFn: () =>
-      api
-        .get('/video/credits', { params: { session_token: sessionToken } })
-        .then((r) => r.data.detail as { balance: number }),
-    enabled: !!sessionToken,
-  })
-  const videoCredits = videoCreditData?.balance ?? 0
-
-  const buyVideoCreditsMutation = useMutation({
-    mutationFn: async (pkg: string) => {
-      const r = await api.post('/video/credits/checkout', {
-        session_token: sessionToken,
-        package: pkg,
-        success_url: `${window.location.origin}/account?video_credits=purchased`,
-        cancel_url: `${window.location.origin}/account`,
-      })
-      return r.data.detail.checkout_url as string
-    },
-    onSuccess: (url) => { window.location.href = url },
-  })
-
-  useEffect(() => {
-    if (tzData?.timezone && !tzInitialised) {
-      setTimezone(tzData.timezone)
-      setTzInitialised(true)
-    }
-  }, [tzData, tzInitialised])
-
-  // Seed blog/sitemap from DB on first load — DB is source of truth over localStorage
-  useEffect(() => {
-    if (settingsData && !urlsInitialised) {
-      const apiBlogUrl = settingsData.blog_url ?? ''
-      const apiSitemapUrl = settingsData.sitemap_url ?? ''
-      if (apiBlogUrl) {
-        setBlogUrl(apiBlogUrl)
-        localStorage.setItem('lem_blog_url', apiBlogUrl)
-      }
-      if (apiSitemapUrl) {
-        setSitemapUrl(apiSitemapUrl)
-        localStorage.setItem('lem_sitemap_url', apiSitemapUrl)
-      }
-      if (settingsData.company_linked_in_url) {
-        setCompanyPage(settingsData.company_linked_in_url)
-      }
-      setUrlsInitialised(true)
-    }
-  }, [settingsData, urlsInitialised])
-
-  const subscription = settingsData?.subscription
-  const tier = subscription?.tier ?? 'free_trial'
-  const trialDays = daysUntil(subscription?.trial_ends_at)
-  const isOnTrial = subscription?.status === 'trial'
-  const isPaidPlan = subscription?.status === 'active'
-  const hasStripeCustomer = !!subscription?.stripe_customer_id
-
-  const hasValidToken = tokenStatusData ? !tokenStatusData.is_expired : false
-  const isLinkedInConnected = liConnectedLocal || hasValidToken
-  const tokenExpiringSoon = tokenStatusData?.is_expiring_soon ?? false
-  const tokenExpired = tokenStatusData?.is_expired ?? false
-
-  const showLinkedInSection = !isLinkedInConnected || tokenExpiringSoon || tokenExpired
-
-  // Profile save
-  const saveMutation = useMutation({
-    mutationFn: () =>
-      api.put('/user/', { email, blog_url: blogUrl || null, sitemap_url: sitemapUrl || null }),
-    onSuccess: () => {
-      localStorage.setItem('lem_blog_url', blogUrl)
-      localStorage.setItem('lem_sitemap_url', sitemapUrl)
-      setSavedMsg('Settings saved!')
-      setTimeout(() => setSavedMsg(null), 3000)
-    },
-    onError: () => {
-      setSavedMsg('Save failed — please try again.')
-      setTimeout(() => setSavedMsg(null), 5000)
-    },
-  })
-
-  // Preferences save
-  const prefsMutation = useMutation({
-    mutationFn: () =>
-      api.put('/user/settings', {
-        session_token: sessionToken,
-        last_login_inactivate_delay: inactivateDelay,
-        auto_schedule_posts: autoSchedule,
-      }),
-    onSuccess: () => {
-      setPrefsSavedMsg('Preferences saved!')
-      setTimeout(() => setPrefsSavedMsg(null), 3000)
-    },
-    onError: () => {
-      setPrefsSavedMsg('Save failed — please try again.')
-      setTimeout(() => setPrefsSavedMsg(null), 5000)
-    },
-  })
-
-  const tzMutation = useMutation({
-    mutationFn: () =>
-      api.put('/user/timezone', { session_token: sessionToken, timezone }),
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ['user-timezone'] })
-      setTzSavedMsg('Timezone saved!')
-      setTimeout(() => setTzSavedMsg(null), 3000)
-    },
-    onError: () => {
-      setTzSavedMsg('Save failed — please try again.')
-      setTimeout(() => setTzSavedMsg(null), 5000)
-    },
-  })
-
-  // Login location — used so the automation browser appears to log in from where
-  // you normally do, reducing LinkedIn "new location" challenges.
-  const { data: locationData } = useQuery({
-    queryKey: ['user-location', sessionToken],
-    queryFn: () =>
-      api
-        .get(`/user/location?session_token=${encodeURIComponent(sessionToken!)}`)
-        .then((r) => r.data.detail as {
-          latitude?: number; longitude?: number; city?: string; country?: string; timezone?: string
-        }),
-    enabled: !!sessionToken,
-  })
-
-  const autocaptureLocationMutation = useMutation({
-    mutationFn: () =>
-      api
-        .post('/user/location/autocapture', { session_token: sessionToken })
-        .then((r) => r.data.detail),
-    onSuccess: (d: { city?: string; country?: string }) => {
-      queryClient.invalidateQueries({ queryKey: ['user-location'] })
-      queryClient.invalidateQueries({ queryKey: ['user-timezone'] })
-      const where = [d?.city, d?.country].filter(Boolean).join(', ')
-      setLocationMsg({ ok: true, text: where ? `Location set to ${where}.` : 'Location detected.' })
-      setTimeout(() => setLocationMsg(null), 4000)
-    },
-    onError: () => {
-      setLocationMsg({ ok: false, text: 'Could not detect your location — try again.' })
-      setTimeout(() => setLocationMsg(null), 5000)
-    },
-  })
-
-  const setLocationByCityMutation = useMutation({
-    mutationFn: () =>
-      api
-        .post('/user/location/by-city', {
-          session_token: sessionToken,
-          city: locCity.trim(),
-          state: locState.trim() || null,
-          country: locCountry.trim() || null,
-        })
-        .then((r) => r.data.detail),
-    onSuccess: (d: { city?: string; country?: string; timezone?: string }) => {
-      queryClient.invalidateQueries({ queryKey: ['user-location'] })
-      queryClient.invalidateQueries({ queryKey: ['user-timezone'] })
-      const where = [d?.city, d?.country].filter(Boolean).join(', ')
-      setLocationMsg({ ok: true, text: `Location set to ${where}${d?.timezone ? ` (${d.timezone})` : ''}.` })
-      setTimeout(() => setLocationMsg(null), 4000)
-    },
-    onError: (e: any) => {
-      const detail = e?.response?.data?.detail
-      setLocationMsg({ ok: false, text: typeof detail === 'string' ? detail : 'Could not set that location — check the city/state.' })
-      setTimeout(() => setLocationMsg(null), 5000)
-    },
-  })
-
-  // Engagement preferences — targeting + voice for AI comments/DMs
-  const { data: engData } = useQuery({
-    queryKey: ['engagement-preferences', sessionToken],
-    queryFn: () =>
-      api
-        .get(`/user/engagement-preferences?session_token=${encodeURIComponent(sessionToken!)}`)
-        .then((r) => r.data.detail as EngPrefs),
-    enabled: !!sessionToken,
-    staleTime: 60 * 1000,
-  })
-  useEffect(() => {
-    if (engData && !engPrefs) setEngPrefs(engData)
-  }, [engData])
-
-  const setEng = (patch: Partial<EngPrefs>) => setEngPrefs((p) => (p ? { ...p, ...patch } : p))
-
-  const engMutation = useMutation({
-    mutationFn: () => api.put('/user/engagement-preferences', { session_token: sessionToken, ...engPrefs }),
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ['engagement-preferences'] })
-      setEngMsg({ ok: true, text: 'Saved.' })
-      setTimeout(() => setEngMsg(null), 3000)
-    },
-    onError: () => {
-      setEngMsg({ ok: false, text: 'Could not save — try again.' })
-      setTimeout(() => setEngMsg(null), 5000)
-    },
-  })
-
-  // DM templates (per event, with optional follow-up steps)
-  const { data: dmData } = useQuery({
-    queryKey: ['dm-templates', sessionToken],
-    queryFn: () =>
-      api
-        .get(`/user/dm-templates?session_token=${encodeURIComponent(sessionToken!)}`)
-        .then((r) => r.data.detail as DmTemplate[]),
-    enabled: !!sessionToken,
-    staleTime: 60 * 1000,
-  })
-  useEffect(() => {
-    if (dmData && !dmInit) {
-      const seeded: DmTemplate[] = [...dmData]
-      for (const ev of DM_EVENTS) {
-        if (!seeded.some((t) => t.event_type === ev.key && t.step === 0)) {
-          seeded.push({ event_type: ev.key, step: 0, delay_hours: 0, template_text: '', is_active: true })
-        }
-      }
-      setDmTemplates(seeded)
-      setDmInit(true)
-    }
-  }, [dmData, dmInit])
-
-  const updateTemplate = (event_type: string, step: number, patch: Partial<DmTemplate>) =>
-    setDmTemplates((ts) => ts.map((t) => (t.event_type === event_type && t.step === step ? { ...t, ...patch } : t)))
-  const addFollowupStep = (event_type: string) =>
-    setDmTemplates((ts) => {
-      const nextStep = ts.filter((t) => t.event_type === event_type).reduce((m, t) => Math.max(m, t.step), -1) + 1
-      return [...ts, { event_type, step: nextStep, delay_hours: 24, template_text: '', is_active: true }]
-    })
-  const removeStep = (event_type: string, step: number) =>
-    setDmTemplates((ts) => ts.filter((t) => !(t.event_type === event_type && t.step === step)))
-
-  const dmMutation = useMutation({
-    mutationFn: () =>
-      api.put('/user/dm-templates', {
-        session_token: sessionToken,
-        templates: dmTemplates.filter((t) => t.template_text.trim()),
-      }),
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ['dm-templates'] })
-      setDmMsg({ ok: true, text: 'DM templates saved.' })
-      setTimeout(() => setDmMsg(null), 3000)
-    },
-    onError: () => {
-      setDmMsg({ ok: false, text: 'Could not save — try again.' })
-      setTimeout(() => setDmMsg(null), 5000)
-    },
-  })
-
-  // Newsletter settings
-  const { data: nlData } = useQuery({
-    queryKey: ['newsletter-settings', sessionToken],
-    queryFn: () =>
-      api
-        .get(`/user/newsletter-settings?session_token=${encodeURIComponent(sessionToken!)}`)
-        .then((r) => r.data.detail as NewsletterSettings),
-    enabled: !!sessionToken,
-    staleTime: 60 * 1000,
-  })
-
-  // LinkedIn Groups — per-group on/off for LEM engagement
-  const { data: groupsData } = useQuery({
-    queryKey: ['user-groups', sessionToken],
-    queryFn: () =>
-      api
-        .get(`/user/groups?session_token=${encodeURIComponent(sessionToken!)}`)
-        .then((r) => r.data.detail as UserGroup[]),
-    enabled: !!sessionToken,
-    staleTime: 60 * 1000,
-  })
-  useEffect(() => {
-    if (nlData && !newsletter) setNewsletter(nlData)
-  }, [nlData])
-
-  const setNl = (patch: Partial<NewsletterSettings>) => setNewsletter((p) => (p ? { ...p, ...patch } : p))
-
-  const nlMutation = useMutation({
-    mutationFn: () => api.put('/user/newsletter-settings', { session_token: sessionToken, ...newsletter }),
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ['newsletter-settings'] })
-      setNlMsg({ ok: true, text: 'Saved.' })
-      setTimeout(() => setNlMsg(null), 3000)
-    },
-    onError: () => {
-      setNlMsg({ ok: false, text: 'Could not save — try again.' })
-      setTimeout(() => setNlMsg(null), 5000)
-    },
-  })
-
-  useEffect(() => {
-    if (groupsData && !groupsInit) {
-      setGroups(groupsData)
-      setGroupsInit(true)
-    }
-  }, [groupsData, groupsInit])
-
-  const toggleGroup = (gid: string) =>
-    setGroups((gs) => gs.map((g) => (g.group_id === gid ? { ...g, enabled: !g.enabled } : g)))
-
-  const groupsMutation = useMutation({
-    mutationFn: () =>
-      api.put('/user/groups', {
-        session_token: sessionToken,
-        groups: Object.fromEntries(groups.map((g) => [g.group_id, g.enabled])),
-      }),
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ['user-groups'] })
-      setGroupsMsg({ ok: true, text: 'Saved.' })
-      setTimeout(() => setGroupsMsg(null), 3000)
-    },
-    onError: () => {
-      setGroupsMsg({ ok: false, text: 'Could not save — try again.' })
-      setTimeout(() => setGroupsMsg(null), 5000)
-    },
-  })
-
-  // Lead magnet (comment→DM)
-  const { data: lmData } = useQuery({
-    queryKey: ['lead-magnet', sessionToken],
-    queryFn: () =>
-      api.get(`/user/lead-magnet?session_token=${encodeURIComponent(sessionToken!)}`).then((r) => r.data.detail as LeadMagnet),
-    enabled: !!sessionToken,
-    staleTime: 60 * 1000,
-  })
-  useEffect(() => {
-    if (lmData && !leadMagnet) setLeadMagnet(lmData)
-  }, [lmData])
-  const setLm = (patch: Partial<LeadMagnet>) => setLeadMagnet((p) => (p ? { ...p, ...patch } : p))
-  const lmMutation = useMutation({
-    mutationFn: () => api.put('/user/lead-magnet', { session_token: sessionToken, ...leadMagnet }),
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ['lead-magnet'] })
-      setLmMsg({ ok: true, text: 'Saved.' }); setTimeout(() => setLmMsg(null), 3000)
-    },
-    onError: () => { setLmMsg({ ok: false, text: 'Could not save — try again.' }); setTimeout(() => setLmMsg(null), 5000) },
-  })
-
-  // Personalized best-times-to-post recommendations (read-only)
-  const { data: postStats } = useQuery({
-    queryKey: ['post-stats', sessionToken],
-    queryFn: () =>
-      api
-        .get(`/user/post-stats?session_token=${encodeURIComponent(sessionToken!)}`)
-        .then((r) => r.data.detail as {
-          recommendations: { weekday: string; hour: number; avg_engagement: number; sample: number }[]
-          sample_size: number
-        }),
-    enabled: !!sessionToken,
-    staleTime: 5 * 60 * 1000,
-  })
-
-  // Stripe checkout redirect
-  const checkoutMutation = useMutation({
-    mutationFn: (tier: string) =>
-      api
-        .post('/billing/create-checkout-session', {
-          session_token: sessionToken,
-          tier,
-          success_url: `${window.location.origin}/account?upgraded=1`,
-          cancel_url: `${window.location.origin}/account`,
-        })
-        .then((r) => r.data.detail as { checkout_url: string | null; upgraded?: boolean }),
-    onSuccess: (detail) => {
-      if (detail.upgraded || !detail.checkout_url) {
-        // In-place upgrade — no Stripe redirect needed. Refresh subscription data.
-        setCheckoutMsg({ ok: true, text: 'Plan updated successfully!' })
-        queryClient.invalidateQueries({ queryKey: ['userSettings'] })
-        setTimeout(() => setCheckoutMsg(null), 5000)
-      } else {
-        window.location.href = detail.checkout_url
-      }
-    },
-    onError: () => {
-      setCheckoutMsg({ ok: false, text: 'Could not start checkout — please try again.' })
-      setTimeout(() => setCheckoutMsg(null), 6000)
-    },
-  })
-
-  // LinkedIn password save
-  const liPasswordMutation = useMutation({
-    mutationFn: () =>
-      api.put('/user/linkedin-password', {
-        session_token: sessionToken,
-        linkedin_password: liPassword,
-      }),
-    onSuccess: () => {
-      setLiPassword('')
-      setLiPasswordMsg({ ok: true, text: 'LinkedIn password saved.' })
-      setTimeout(() => setLiPasswordMsg(null), 4000)
-    },
-    onError: () => {
-      setLiPasswordMsg({ ok: false, text: 'Save failed — please try again.' })
-      setTimeout(() => setLiPasswordMsg(null), 5000)
-    },
-  })
-
-  // LinkedIn company page save (used by the monthly company-page invite automation)
-  const companyPageMutation = useMutation({
-    mutationFn: () =>
-      api.put('/user/company-page', {
-        session_token: sessionToken,
-        company_linked_in_url: companyPage.trim() || null,
-      }),
-    onSuccess: () => {
-      setCompanyPageMsg({ ok: true, text: companyPage.trim() ? 'Company page saved.' : 'Company page cleared.' })
-      setTimeout(() => setCompanyPageMsg(null), 4000)
-    },
-    onError: () => {
-      setCompanyPageMsg({ ok: false, text: 'Save failed — use a full linkedin.com/company/… URL.' })
-      setTimeout(() => setCompanyPageMsg(null), 6000)
-    },
-  })
-
-  // Stripe portal redirect
-  const portalMutation = useMutation({
-    mutationFn: () =>
-      api
-        .post('/billing/create-portal-session', {
-          session_token: sessionToken,
-          return_url: `${window.location.origin}/account`,
-        })
-        .then((r) => r.data.detail.portal_url as string),
-    onSuccess: (url) => { window.location.href = url },
-    onError: () => setSavedMsg('Could not open billing portal — please try again.'),
-  })
-
-  function handleSave(e: React.FormEvent) {
-    e.preventDefault()
-    saveMutation.mutate()
-  }
-
-  function handlePrefsSave(e: React.FormEvent) {
-    e.preventDefault()
-    prefsMutation.mutate()
-  }
 
   return (
     <div className="max-w-2xl space-y-6">
       <h1 className="text-2xl font-bold text-gray-800">Account Settings</h1>
+
+      {oauthMsg && (
+        <p className={`text-sm font-medium ${oauthMsg.ok ? 'text-green-600' : 'text-red-600'}`}>
+          {oauthMsg.text}
+        </p>
+      )}
 
       {/* Account setup checklist — exactly what automation needs, from the readiness API */}
       {readiness && !readiness.ready && (
@@ -744,775 +99,57 @@ export default function Account() {
         </div>
       )}
 
-      <SectionHeading title="Plan & Billing" />
-
-      {/* Subscription card */}
-      <div className="bg-white rounded-lg shadow-sm border border-gray-200 p-6 space-y-4">
-        <div className="flex items-center justify-between">
-          <h2 className="text-base font-semibold text-gray-700">
-            Subscription <span className="text-red-500">*</span>
-          </h2>
-          <span className={`text-xs font-semibold px-2.5 py-1 rounded-full ${TIER_COLORS[tier] ?? TIER_COLORS.free_trial}`}>
-            {TIER_LABELS[tier] ?? tier}
-          </span>
-        </div>
-
-        {isOnTrial && trialDays !== null && (
-          <div className={`rounded-lg p-3 text-sm ${trialDays <= 3 ? 'bg-red-50 text-red-800 border border-red-200' : 'bg-yellow-50 text-yellow-800 border border-yellow-200'}`}>
-            {trialDays === 0
-              ? 'Your free trial has expired. Upgrade to keep using LEM.'
-              : `Free trial: ${trialDays} day${trialDays === 1 ? '' : 's'} remaining.`}
-          </div>
-        )}
-
-        {isPaidPlan && (
-          <p className="text-sm text-green-700">Your subscription is active.</p>
-        )}
-
-        {/* Upgrade options — shown to trial users without a paid plan */}
-        {!isPaidPlan && hasStripeCustomer && (
-          <div className="space-y-2">
-            <p className="text-xs text-gray-500 font-medium uppercase tracking-wide">Upgrade your plan</p>
-            <div className="grid grid-cols-3 gap-2">
-              {[
-                { tier: 'starter', label: 'Starter', price: '$29/mo' },
-                { tier: 'professional', label: 'Professional', price: '$79/mo' },
-                { tier: 'enterprise', label: 'Enterprise', price: '$199/mo' },
-              ].map(({ tier: t, label, price }) => (
-                <button
-                  key={t}
-                  onClick={() => checkoutMutation.mutate(t)}
-                  disabled={checkoutMutation.isPending}
-                  className="flex flex-col items-center border border-blue-200 rounded-lg p-3 hover:bg-blue-50 transition-colors disabled:opacity-50 text-center"
-                >
-                  <span className="text-sm font-semibold text-blue-800">{label}</span>
-                  <span className="text-xs text-gray-500">{price}</span>
-                </button>
-              ))}
-            </div>
-            {checkoutMutation.isPending && (
-              <p className="text-xs text-blue-600">Redirecting to checkout…</p>
-            )}
-            {checkoutMsg && (
-              <p className={`text-sm font-medium ${checkoutMsg.ok ? 'text-green-600' : 'text-red-600'}`}>
-                {checkoutMsg.text}
-              </p>
-            )}
-          </div>
-        )}
-
-        {/* Billing portal — shown to paid subscribers */}
-        {isPaidPlan && hasStripeCustomer && (
+      {/* Tab switcher */}
+      <div className="flex flex-wrap gap-1 border-b border-gray-200">
+        {TABS.map((t) => (
           <button
-            onClick={() => portalMutation.mutate()}
-            disabled={portalMutation.isPending}
-            className="text-sm text-blue-600 hover:underline disabled:opacity-50"
-          >
-            Manage billing →
-          </button>
-        )}
-
-        {!hasStripeCustomer && (
-          <p className="text-xs text-gray-400">
-            Billing not yet configured for your account. Contact support if you need help.
-          </p>
-        )}
-      </div>
-
-      {/* Premium video credits */}
-      <div className="bg-white rounded-lg shadow-sm border border-gray-200 p-6 space-y-4">
-        <div className="flex items-center justify-between">
-          <h2 className="text-base font-semibold text-gray-700">Premium Video Credits</h2>
-          <span className="text-2xl font-bold text-blue-900" data-testid="video-credit-balance">{videoCredits}</span>
-        </div>
-        <p className="text-sm text-gray-500">
-          Premium videos use Google Veo (realistic motion + native audio): 1 credit per premium video,
-          3 for top realism. Standard videos are always free with your plan.
-        </p>
-        <div className="grid grid-cols-2 sm:grid-cols-4 gap-3">
-          {VIDEO_PACKAGES.map((pkg) => (
-            <div
-              key={pkg.key}
-              className={`relative border rounded-xl p-4 flex flex-col items-center gap-1 ${
-                pkg.key === 'medium' ? 'border-blue-400 bg-blue-50' : 'border-gray-200 bg-white'
-              }`}
-            >
-              <span className="text-xl font-bold text-gray-900">{pkg.price}</span>
-              <span className="text-xs font-medium text-gray-600 text-center">{pkg.label}</span>
-              <button
-                onClick={() => buyVideoCreditsMutation.mutate(pkg.key)}
-                disabled={buyVideoCreditsMutation.isPending}
-                className="mt-2 w-full py-1.5 rounded-lg bg-blue-600 hover:bg-blue-700 text-white text-sm font-medium disabled:opacity-50 transition-colors"
-              >
-                {buyVideoCreditsMutation.isPending ? '…' : 'Buy'}
-              </button>
-            </div>
-          ))}
-        </div>
-      </div>
-
-      <SectionHeading title="LinkedIn" />
-
-      {/* LinkedIn connection — hidden when connected and token healthy */}
-      {showLinkedInSection && (
-        <div className="bg-white rounded-lg shadow-sm border border-gray-200 p-6 space-y-4">
-          <h2 className="text-base font-semibold text-gray-700">
-            LinkedIn Connection <span className="text-red-500">*</span>
-            <span className="ml-2 text-[11px] font-semibold text-red-600 bg-red-50 px-2 py-0.5 rounded">
-              Required
-            </span>
-          </h2>
-
-          {isLinkedInConnected && tokenExpiringSoon && !tokenExpired && (
-            <div className="bg-yellow-50 border border-yellow-200 rounded-lg p-3 text-sm text-yellow-800">
-              Your LinkedIn token is expiring soon. Please reconnect to keep posting.
-              {tokenStatusData?.refresh_attempted && !tokenStatusData?.refresh_succeeded && (
-                <span className="ml-1 text-xs">(Auto-refresh failed — manual reconnect required.)</span>
-              )}
-            </div>
-          )}
-
-          {tokenExpired && isLinkedInConnected && (
-            <div className="bg-red-50 border border-red-200 rounded-lg p-3 text-sm text-red-800">
-              Your LinkedIn connection has expired. Please reconnect to continue posting.
-            </div>
-          )}
-
-          <div className="flex items-center gap-3">
-            <div
-              className={`w-3 h-3 rounded-full flex-shrink-0 ${
-                isLinkedInConnected && !tokenExpired ? 'bg-green-500' : 'bg-gray-300'
-              }`}
-            />
-            <span className="text-sm text-gray-600">
-              {isLinkedInConnected && !tokenExpired ? 'Connected to LinkedIn' : 'Not connected'}
-            </span>
-          </div>
-
-          <p className="text-xs text-gray-500">
-            Connecting allows LEM to post on your behalf, send DMs, reply to comments, and engage with your network automatically.
-          </p>
-
-          <a
-            href={`${LI_AUTH_URL}?email=${encodeURIComponent(email)}&session_token=${encodeURIComponent(sessionToken ?? '')}`}
-            className="inline-block w-full text-center py-2 rounded-lg text-sm font-semibold bg-blue-700 text-white hover:bg-blue-800 cursor-pointer transition-colors"
-          >
-            {isLinkedInConnected ? 'Reconnect LinkedIn' : 'Connect LinkedIn'}
-          </a>
-        </div>
-      )}
-
-      {/* LinkedIn Session (cookie) — preferred login method; avoids the device challenge */}
-      <LinkedInSessionCard connected={sessionOk} />
-
-      {/* LinkedIn Company Page — drives the monthly connection-invite automation */}
-      <div className="bg-white rounded-lg shadow-sm border border-gray-200 p-6 space-y-4">
-        <div>
-          <h2 className="text-base font-semibold text-gray-700">LinkedIn Company Page</h2>
-          <p className="text-xs text-gray-500 mt-1">
-            If you have a LinkedIn company page, add it here. On the 1st of each month LEM
-            invites your connections to follow it (using your monthly invite credits, which
-            reset monthly). Leave blank if you don't have one.
-          </p>
-        </div>
-
-        <div>
-          <label className="block text-sm font-medium text-gray-700 mb-1">Company page URL</label>
-          <input
-            type="url"
-            value={companyPage}
-            onChange={(e) => setCompanyPage(e.target.value)}
-            className="w-full border border-gray-300 rounded-lg px-3 py-2 text-sm focus:outline-none focus:ring-2 focus:ring-blue-500"
-            placeholder="https://www.linkedin.com/company/your-company/"
-          />
-        </div>
-
-        {companyPageMsg && (
-          <p className={`text-sm font-medium ${companyPageMsg.ok ? 'text-green-600' : 'text-red-600'}`}>
-            {companyPageMsg.text}
-          </p>
-        )}
-
-        <button
-          type="button"
-          onClick={() => companyPageMutation.mutate()}
-          disabled={companyPageMutation.isPending}
-          className="w-full bg-blue-600 text-white py-2 rounded-lg text-sm font-semibold hover:bg-blue-700 disabled:opacity-50 transition-colors"
-        >
-          {companyPageMutation.isPending ? 'Saving…' : 'Save Company Page'}
-        </button>
-      </div>
-
-      {/* LinkedIn Automation Password — always shown when LinkedIn is connected */}
-      {isLinkedInConnected && !tokenExpired && (
-        <div className="bg-white rounded-lg shadow-sm border border-gray-200 p-6 space-y-4">
-          <div>
-            <h2 className="text-base font-semibold text-gray-700">LinkedIn Automation Password</h2>
-            <p className="text-xs text-gray-500 mt-1">
-              LEM uses your LinkedIn password to log in via browser automation for actions like
-              profile scraping, post personalization, and DM engagement. It is stored securely and
-              never shared or displayed after saving.
-            </p>
-          </div>
-
-          <div>
-            <label className="block text-sm font-medium text-gray-700 mb-1">LinkedIn Password</label>
-            <div className="relative">
-              <input
-                type={showLiPassword ? 'text' : 'password'}
-                value={liPassword}
-                onChange={(e) => setLiPassword(e.target.value)}
-                className="w-full border border-gray-300 rounded-lg px-3 py-2 text-sm pr-16 focus:outline-none focus:ring-2 focus:ring-blue-500"
-                placeholder="Enter your LinkedIn password"
-                autoComplete="new-password"
-              />
-              <button
-                type="button"
-                onClick={() => setShowLiPassword((v) => !v)}
-                className="absolute inset-y-0 right-2 flex items-center px-2 text-xs text-gray-500 hover:text-gray-700"
-              >
-                {showLiPassword ? 'Hide' : 'Show'}
-              </button>
-            </div>
-          </div>
-
-          {liPasswordMsg && (
-            <p className={`text-sm font-medium ${liPasswordMsg.ok ? 'text-green-600' : 'text-red-600'}`}>
-              {liPasswordMsg.text}
-            </p>
-          )}
-
-          <button
+            key={t.key}
             type="button"
-            onClick={() => liPasswordMutation.mutate()}
-            disabled={liPasswordMutation.isPending || !liPassword}
-            className="w-full bg-blue-600 text-white py-2 rounded-lg text-sm font-semibold hover:bg-blue-700 disabled:opacity-50 transition-colors"
+            onClick={() => setActiveTab(t.key)}
+            className={`px-3 py-2 text-sm font-semibold border-b-2 -mb-px transition-colors ${
+              activeTab === t.key
+                ? 'border-blue-600 text-blue-700'
+                : 'border-transparent text-gray-500 hover:text-gray-700 hover:border-gray-300'
+            }`}
           >
-            {liPasswordMutation.isPending ? 'Saving…' : 'Save Password'}
+            {t.label}
           </button>
-        </div>
-      )}
-
-      <SectionHeading title="Content & Profile" />
-
-      {/* Email + content sources the AI uses for post ideas */}
-      <form onSubmit={handleSave} className="bg-white rounded-lg shadow-sm border border-gray-200 p-6 space-y-4">
-        <h2 className="text-base font-semibold text-gray-700">Content &amp; Profile</h2>
-
-        <div>
-          <label className="block text-sm font-medium text-gray-700 mb-1">Email</label>
-          <p className="text-sm text-gray-800 bg-gray-50 border border-gray-200 rounded-lg px-3 py-2">
-            {email}
-          </p>
-          <p className="text-xs text-green-600 mt-1">✓ Verified email</p>
-        </div>
-
-        <div>
-          <label className="block text-sm font-medium text-gray-700 mb-1">Blog URL</label>
-          <input
-            type="url"
-            value={blogUrl}
-            onChange={(e) => setBlogUrl(e.target.value)}
-            className="w-full border border-gray-300 rounded-lg px-3 py-2 text-sm focus:outline-none focus:ring-2 focus:ring-blue-500"
-            placeholder="https://yourblog.com"
-          />
-        </div>
-
-        <div>
-          <label className="block text-sm font-medium text-gray-700 mb-1">Sitemap URL</label>
-          <input
-            type="url"
-            value={sitemapUrl}
-            onChange={(e) => setSitemapUrl(e.target.value)}
-            className="w-full border border-gray-300 rounded-lg px-3 py-2 text-sm focus:outline-none focus:ring-2 focus:ring-blue-500"
-            placeholder="https://yourblog.com/sitemap.xml"
-          />
-          <p className="text-xs text-gray-400 mt-1">Used by AI to generate content ideas from your existing posts.</p>
-        </div>
-
-        {savedMsg && (
-          <p className={`text-sm font-medium ${saveMutation.isError ? 'text-red-600' : 'text-green-600'}`}>
-            {savedMsg}
-          </p>
-        )}
-
-        <button
-          type="submit"
-          disabled={saveMutation.isPending}
-          className="w-full bg-blue-600 text-white py-2 rounded-lg text-sm font-semibold hover:bg-blue-700 disabled:opacity-50 transition-colors"
-        >
-          {saveMutation.isPending ? 'Saving…' : 'Save'}
-        </button>
-      </form>
-
-      <SectionHeading title="Automation Settings" />
-
-      {/* Preferences card */}
-      <form onSubmit={handlePrefsSave} className="bg-white rounded-lg shadow-sm border border-gray-200 p-6 space-y-5">
-        <h2 className="text-base font-semibold text-gray-700">Preferences</h2>
-
-        <div>
-          <label className="block text-sm font-medium text-gray-700 mb-1">
-            Inactivity auto-stop delay
-          </label>
-          <select
-            value={inactivateDelay === null ? 'never' : String(inactivateDelay)}
-            onChange={(e) =>
-              setInactivateDelay(e.target.value === 'never' ? null : Number(e.target.value))
-            }
-            className="w-full border border-gray-300 rounded-lg px-3 py-2 text-sm focus:outline-none focus:ring-2 focus:ring-blue-500"
-          >
-            {INACTIVATE_OPTIONS.map(({ label, value }) => (
-              <option key={String(value)} value={value === null ? 'never' : String(value)}>
-                {label}
-              </option>
-            ))}
-          </select>
-          <p className="text-xs text-gray-400 mt-1">
-            If you haven't logged in within this window, LEM will pause automated posting and LinkedIn activity to avoid acting on your behalf unintentionally.
-            Set to "Never" to keep automation running regardless of login activity.
-          </p>
-        </div>
-
-        <div>
-          <div className="flex items-center justify-between">
-            <div>
-              <p className="text-sm font-medium text-gray-700">Auto-schedule AI posts</p>
-              <p className="text-xs text-gray-400 mt-0.5">
-                When on, AI-generated posts are automatically approved and queued for posting.
-                When off, each post waits for your manual approval in the Review tab.
-              </p>
-            </div>
-            <button
-              type="button"
-              onClick={() => setAutoSchedule((v) => !v)}
-              className={`relative inline-flex h-6 w-11 flex-shrink-0 cursor-pointer rounded-full border-2 border-transparent transition-colors duration-200 focus:outline-none focus:ring-2 focus:ring-blue-500 focus:ring-offset-2 ${
-                autoSchedule ? 'bg-blue-600' : 'bg-gray-200'
-              }`}
-              role="switch"
-              aria-checked={autoSchedule}
-            >
-              <span
-                className={`pointer-events-none inline-block h-5 w-5 transform rounded-full bg-white shadow ring-0 transition duration-200 ease-in-out ${
-                  autoSchedule ? 'translate-x-5' : 'translate-x-0'
-                }`}
-              />
-            </button>
-          </div>
-        </div>
-
-        {prefsSavedMsg && (
-          <p className={`text-sm font-medium ${prefsMutation.isError ? 'text-red-600' : 'text-green-600'}`}>
-            {prefsSavedMsg}
-          </p>
-        )}
-
-        <button
-          type="submit"
-          disabled={prefsMutation.isPending}
-          className="w-full bg-blue-600 text-white py-2 rounded-lg text-sm font-semibold hover:bg-blue-700 disabled:opacity-50 transition-colors"
-        >
-          {prefsMutation.isPending ? 'Saving…' : 'Save Preferences'}
-        </button>
-      </form>
-
-      {/* Voice & Tone card */}
-      {engPrefs && (
-        <div className="bg-white rounded-lg shadow-sm border border-gray-200 p-6 space-y-5">
-          <h2 className="text-base font-semibold text-gray-700">Voice &amp; Tone</h2>
-          <p className="text-xs text-gray-500">
-            How AI comments and DMs should sound. Leave blank to infer purely from your profile.
-          </p>
-          <div className="grid grid-cols-1 sm:grid-cols-2 gap-4">
-            <div>
-              <label className="block text-sm font-medium text-gray-700 mb-1">Tone</label>
-              <input type="text" value={engPrefs.tone || ''} onChange={(e) => setEng({ tone: e.target.value })}
-                placeholder="e.g. warm, authoritative"
-                className="w-full border border-gray-300 rounded-lg px-3 py-2 text-sm" />
-            </div>
-            <div>
-              <label className="block text-sm font-medium text-gray-700 mb-1">Comment length</label>
-              <select value={engPrefs.comment_length} onChange={(e) => setEng({ comment_length: e.target.value })}
-                className="w-full border border-gray-300 rounded-lg px-3 py-2 text-sm">
-                <option value="short">Short (~300 chars)</option>
-                <option value="medium">Medium (~600 chars)</option>
-                <option value="long">Long (~1100 chars)</option>
-              </select>
-            </div>
-          </div>
-          <div>
-            <label className="block text-sm font-medium text-gray-700 mb-1">Style guidance</label>
-            <input type="text" value={engPrefs.comment_style || ''} onChange={(e) => setEng({ comment_style: e.target.value })}
-              placeholder="e.g. ask a question, avoid buzzwords"
-              className="w-full border border-gray-300 rounded-lg px-3 py-2 text-sm" />
-          </div>
-          <div className="flex items-center justify-between">
-            <p className="text-sm font-medium text-gray-700">Use emojis</p>
-            <Toggle on={engPrefs.use_emojis} onClick={() => setEng({ use_emojis: !engPrefs.use_emojis })} />
-          </div>
-          <div className="flex items-center justify-between">
-            <p className="text-sm font-medium text-gray-700">Use hashtags</p>
-            <Toggle on={engPrefs.use_hashtags} onClick={() => setEng({ use_hashtags: !engPrefs.use_hashtags })} />
-          </div>
-          {engMsg && (
-            <p className={`text-sm font-medium ${engMsg.ok ? 'text-green-600' : 'text-red-600'}`}>{engMsg.text}</p>
-          )}
-          <button type="button" onClick={() => engMutation.mutate()} disabled={engMutation.isPending}
-            className="w-full bg-blue-600 text-white py-2 rounded-lg text-sm font-semibold hover:bg-blue-700 disabled:opacity-50 transition-colors">
-            {engMutation.isPending ? 'Saving…' : 'Save Voice & Tone'}
-          </button>
-        </div>
-      )}
-
-      {/* Engagement Targeting card */}
-      {engPrefs && (
-        <div className="bg-white rounded-lg shadow-sm border border-gray-200 p-6 space-y-5">
-          <h2 className="text-base font-semibold text-gray-700">Engagement Targeting</h2>
-          <p className="text-xs text-gray-500">
-            Control which posts LEM comments on. Comma-separated. Exclusions always win; if any
-            "include" is set, a post must match one (keyword/author literal, or a topic via AI relevance).
-          </p>
-          {([
-            ['include_topics', 'Include topics (AI relevance)'],
-            ['exclude_topics', 'Exclude topics'],
-            ['include_keywords', 'Include keywords'],
-            ['exclude_keywords', 'Exclude keywords'],
-            ['include_authors', 'Include authors'],
-            ['exclude_authors', 'Exclude authors'],
-          ] as [keyof EngPrefs, string][]).map(([field, label]) => (
-            <div key={field}>
-              <label className="block text-sm font-medium text-gray-700 mb-1">{label}</label>
-              <input type="text" value={csv(engPrefs[field] as string[])}
-                onChange={(e) => setEng({ [field]: parseCsv(e.target.value) } as Partial<EngPrefs>)}
-                placeholder="comma, separated, values"
-                className="w-full border border-gray-300 rounded-lg px-3 py-2 text-sm" />
-            </div>
-          ))}
-          <div className="grid grid-cols-1 sm:grid-cols-3 gap-4">
-            <div>
-              <label className="block text-sm font-medium text-gray-700 mb-1">Min. reactions</label>
-              <input type="number" min={0} value={engPrefs.min_reactions ?? ''}
-                onChange={(e) => setEng({ min_reactions: e.target.value === '' ? null : Number(e.target.value) })}
-                className="w-full border border-gray-300 rounded-lg px-3 py-2 text-sm" />
-            </div>
-            <div>
-              <label className="block text-sm font-medium text-gray-700 mb-1">Max post age (hrs)</label>
-              <input type="number" min={1} value={engPrefs.max_post_age_hours ?? ''}
-                onChange={(e) => setEng({ max_post_age_hours: e.target.value === '' ? null : Number(e.target.value) })}
-                placeholder="24"
-                className="w-full border border-gray-300 rounded-lg px-3 py-2 text-sm" />
-            </div>
-            <div>
-              <label className="block text-sm font-medium text-gray-700 mb-1">Max comments/day</label>
-              <input type="number" min={0} value={engPrefs.max_comments_per_day}
-                onChange={(e) => setEng({ max_comments_per_day: Number(e.target.value) })}
-                className="w-full border border-gray-300 rounded-lg px-3 py-2 text-sm" />
-            </div>
-            <div>
-              <label className="block text-sm font-medium text-gray-700 mb-1">Max DMs/day</label>
-              <input type="number" min={0} value={engPrefs.max_dms_per_day}
-                onChange={(e) => setEng({ max_dms_per_day: Number(e.target.value) })}
-                className="w-full border border-gray-300 rounded-lg px-3 py-2 text-sm" />
-            </div>
-          </div>
-          <div className="flex items-center justify-between">
-            <p className="text-sm font-medium text-gray-700">Reply to comments on my posts</p>
-            <Toggle on={engPrefs.reply_to_own_comments} onClick={() => setEng({ reply_to_own_comments: !engPrefs.reply_to_own_comments })} />
-          </div>
-          {engMsg && (
-            <p className={`text-sm font-medium ${engMsg.ok ? 'text-green-600' : 'text-red-600'}`}>{engMsg.text}</p>
-          )}
-          <button type="button" onClick={() => engMutation.mutate()} disabled={engMutation.isPending}
-            className="w-full bg-blue-600 text-white py-2 rounded-lg text-sm font-semibold hover:bg-blue-700 disabled:opacity-50 transition-colors">
-            {engMutation.isPending ? 'Saving…' : 'Save Targeting'}
-          </button>
-        </div>
-      )}
-
-      {/* Newsletter card */}
-      {newsletter && (
-        <div className="bg-white rounded-lg shadow-sm border border-gray-200 p-6 space-y-4">
-          <div className="flex items-center justify-between">
-            <div>
-              <h2 className="text-base font-semibold text-gray-700">LinkedIn Newsletter</h2>
-              <p className="text-xs text-gray-500">Auto-publish a recurring newsletter (bypasses the feed — subscribers get a notification + email). Repurposes your blog when set.</p>
-            </div>
-            <Toggle on={newsletter.enabled} onClick={() => setNl({ enabled: !newsletter.enabled })} />
-          </div>
-          {newsletter.enabled && (
-            <>
-              <div className="grid grid-cols-1 sm:grid-cols-2 gap-4">
-                <div>
-                  <label className="block text-sm font-medium text-gray-700 mb-1">Title</label>
-                  <input type="text" value={newsletter.title || ''} onChange={(e) => setNl({ title: e.target.value })}
-                    placeholder="e.g. The Growth Brief" className="w-full border border-gray-300 rounded-lg px-3 py-2 text-sm" />
-                </div>
-                <div>
-                  <label className="block text-sm font-medium text-gray-700 mb-1">Cadence</label>
-                  <select value={newsletter.cadence} onChange={(e) => setNl({ cadence: e.target.value })}
-                    className="w-full border border-gray-300 rounded-lg px-3 py-2 text-sm">
-                    <option value="weekly">Weekly (recommended)</option>
-                    <option value="biweekly">Bi-weekly</option>
-                    <option value="monthly">Monthly</option>
-                  </select>
-                </div>
-              </div>
-              <div>
-                <label className="block text-sm font-medium text-gray-700 mb-1">Topic / theme</label>
-                <input type="text" value={newsletter.topic || ''} onChange={(e) => setNl({ topic: e.target.value })}
-                  placeholder="What each edition should focus on" className="w-full border border-gray-300 rounded-lg px-3 py-2 text-sm" />
-              </div>
-              <div className="flex items-center justify-between">
-                <p className="text-sm font-medium text-gray-700">Align with my blog</p>
-                <Toggle on={newsletter.align_with_blog} onClick={() => setNl({ align_with_blog: !newsletter.align_with_blog })} />
-              </div>
-            </>
-          )}
-          {nlMsg && <p className={`text-sm font-medium ${nlMsg.ok ? 'text-green-600' : 'text-red-600'}`}>{nlMsg.text}</p>}
-          <button type="button" onClick={() => nlMutation.mutate()} disabled={nlMutation.isPending}
-            className="w-full bg-blue-600 text-white py-2 rounded-lg text-sm font-semibold hover:bg-blue-700 disabled:opacity-50 transition-colors">
-            {nlMutation.isPending ? 'Saving…' : 'Save Newsletter Settings'}
-          </button>
-        </div>
-      )}
-
-      {/* Lead magnet card */}
-      {leadMagnet && (
-        <div className="bg-white rounded-lg shadow-sm border border-gray-200 p-6 space-y-4">
-          <div className="flex items-center justify-between">
-            <div>
-              <h2 className="text-base font-semibold text-gray-700">Comment → DM Lead Magnet</h2>
-              <p className="text-xs text-gray-500">When someone comments your keyword on your post, auto-DM them a resource. The compliant way to share links (links in posts are penalized).</p>
-            </div>
-            <Toggle on={leadMagnet.enabled} onClick={() => setLm({ enabled: !leadMagnet.enabled })} />
-          </div>
-          {leadMagnet.enabled && (
-            <>
-              <div>
-                <label className="block text-sm font-medium text-gray-700 mb-1">Trigger keyword</label>
-                <input type="text" value={leadMagnet.keyword || ''} onChange={(e) => setLm({ keyword: e.target.value })}
-                  placeholder="e.g. GUIDE" className="w-full border border-gray-300 rounded-lg px-3 py-2 text-sm" />
-              </div>
-              <div>
-                <label className="block text-sm font-medium text-gray-700 mb-1">DM message (may include your link)</label>
-                <textarea value={leadMagnet.message || ''} onChange={(e) => setLm({ message: e.target.value })} rows={3}
-                  placeholder="Thanks for the interest! Here's the resource: …" className="w-full border border-gray-300 rounded-lg px-3 py-2 text-sm" />
-              </div>
-            </>
-          )}
-          {lmMsg && <p className={`text-sm font-medium ${lmMsg.ok ? 'text-green-600' : 'text-red-600'}`}>{lmMsg.text}</p>}
-          <button type="button" onClick={() => lmMutation.mutate()} disabled={lmMutation.isPending}
-            className="w-full bg-blue-600 text-white py-2 rounded-lg text-sm font-semibold hover:bg-blue-700 disabled:opacity-50 transition-colors">
-            {lmMutation.isPending ? 'Saving…' : 'Save Lead Magnet'}
-          </button>
-        </div>
-      )}
-
-      {/* Best times to post (data-driven) */}
-      {postStats && (
-        <div className="bg-white rounded-lg shadow-sm border border-gray-200 p-6 space-y-3">
-          <h2 className="text-base font-semibold text-gray-700">Your Best Times to Post</h2>
-          {postStats.recommendations.length > 0 ? (
-            <>
-              <p className="text-xs text-gray-500">Learned from your own post engagement — scheduling leans toward these.</p>
-              <ul className="text-sm text-gray-700 space-y-1">
-                {postStats.recommendations.map((r, i) => (
-                  <li key={i} className="flex justify-between">
-                    <span>{r.weekday} @ {String(r.hour).padStart(2, '0')}:00</span>
-                    <span className="text-gray-400">avg engagement {r.avg_engagement} · {r.sample} post(s)</span>
-                  </li>
-                ))}
-              </ul>
-            </>
-          ) : (
-            <p className="text-xs text-gray-500">Gathering data — recommendations appear after a few posts have engagement stats (currently {postStats.sample_size}).</p>
-          )}
-        </div>
-      )}
-
-      {/* Groups card */}
-      {groupsInit && groups.length > 0 && (
-        <div className="bg-white rounded-lg shadow-sm border border-gray-200 p-6 space-y-4">
-          <h2 className="text-base font-semibold text-gray-700">LinkedIn Groups</h2>
-          <p className="text-xs text-gray-500">Choose which of your joined groups LEM engages in (value-add comments + occasional posts). All on by default.</p>
-          <div className="divide-y divide-gray-100">
-            {groups.map((g) => (
-              <div key={g.group_id} className="flex items-center justify-between py-2">
-                <span className="text-sm text-gray-700 truncate pr-3">{g.group_name || `Group ${g.group_id}`}</span>
-                <Toggle on={g.enabled} onClick={() => toggleGroup(g.group_id)} />
-              </div>
-            ))}
-          </div>
-          {groupsMsg && <p className={`text-sm font-medium ${groupsMsg.ok ? 'text-green-600' : 'text-red-600'}`}>{groupsMsg.text}</p>}
-          <button type="button" onClick={() => groupsMutation.mutate()} disabled={groupsMutation.isPending}
-            className="w-full bg-blue-600 text-white py-2 rounded-lg text-sm font-semibold hover:bg-blue-700 disabled:opacity-50 transition-colors">
-            {groupsMutation.isPending ? 'Saving…' : 'Save Group Settings'}
-          </button>
-        </div>
-      )}
-
-      {/* DM Templates card */}
-      {dmInit && (
-        <div className="bg-white rounded-lg shadow-sm border border-gray-200 p-6 space-y-4">
-          <h2 className="text-base font-semibold text-gray-700">DM Templates</h2>
-          <p className="text-xs text-gray-500">
-            Placeholders: <code>{'{first_name}'}</code>, <code>{'{headline}'}</code>, <code>{'{blog_url}'}</code>.
-            Blank uses the built-in default. Add follow-up steps to message again after a delay — the
-            sequence stops automatically if they reply.
-          </p>
-          {DM_EVENTS.map((ev) => {
-            const steps = dmTemplates.filter((t) => t.event_type === ev.key).sort((a, b) => a.step - b.step)
-            return (
-              <div key={ev.key} className="border-t border-gray-100 pt-4">
-                <p className="text-sm font-semibold text-gray-700">{ev.label}</p>
-                {steps.map((t) => (
-                  <div key={t.step} className="mt-2 space-y-1">
-                    <div className="flex items-center gap-2 text-xs text-gray-500">
-                      <span>{t.step === 0 ? 'Initial message' : `Follow-up ${t.step}`}</span>
-                      {t.step > 0 && (
-                        <label className="flex items-center">
-                          after
-                          <input type="number" min={1} value={t.delay_hours}
-                            onChange={(e) => updateTemplate(ev.key, t.step, { delay_hours: Number(e.target.value) })}
-                            className="mx-1 w-16 border border-gray-300 rounded px-1 py-0.5" />
-                          h
-                        </label>
-                      )}
-                      {t.step > 0 && (
-                        <button type="button" onClick={() => removeStep(ev.key, t.step)}
-                          className="ml-auto text-red-500 hover:text-red-600">Remove</button>
-                      )}
-                    </div>
-                    <textarea value={t.template_text} rows={2}
-                      onChange={(e) => updateTemplate(ev.key, t.step, { template_text: e.target.value })}
-                      placeholder="Leave blank for the default message"
-                      className="w-full border border-gray-300 rounded-lg px-3 py-2 text-sm" />
-                  </div>
-                ))}
-                <button type="button" onClick={() => addFollowupStep(ev.key)}
-                  className="mt-2 text-xs text-blue-600 font-medium hover:text-blue-700">+ Add follow-up</button>
-              </div>
-            )
-          })}
-          {dmMsg && (
-            <p className={`text-sm font-medium ${dmMsg.ok ? 'text-green-600' : 'text-red-600'}`}>{dmMsg.text}</p>
-          )}
-          <button type="button" onClick={() => dmMutation.mutate()} disabled={dmMutation.isPending}
-            className="w-full bg-blue-600 text-white py-2 rounded-lg text-sm font-semibold hover:bg-blue-700 disabled:opacity-50 transition-colors">
-            {dmMutation.isPending ? 'Saving…' : 'Save DM Templates'}
-          </button>
-        </div>
-      )}
-
-      {/* Timezone card */}
-      <form
-        onSubmit={(e) => { e.preventDefault(); tzMutation.mutate() }}
-        className="bg-white rounded-lg shadow-sm border border-gray-200 p-6 space-y-4"
-      >
-        <h2 className="text-base font-semibold text-gray-700">Timezone</h2>
-        <div>
-          <label className="block text-sm font-medium text-gray-700 mb-1">Your timezone</label>
-          <select
-            value={timezone}
-            onChange={(e) => setTimezone(e.target.value)}
-            className="w-full border border-gray-300 rounded-lg px-3 py-2 text-sm focus:outline-none focus:ring-2 focus:ring-blue-500"
-          >
-            {COMMON_TIMEZONES.map((tz) => (
-              <option key={tz} value={tz}>{tz}</option>
-            ))}
-          </select>
-          <p className="text-xs text-gray-400 mt-1">
-            Scheduled post times will be displayed in this timezone across all pages.
-          </p>
-        </div>
-
-        {tzSavedMsg && (
-          <p className={`text-sm font-medium ${tzMutation.isError ? 'text-red-600' : 'text-green-600'}`}>
-            {tzSavedMsg}
-          </p>
-        )}
-
-        <button
-          type="submit"
-          disabled={tzMutation.isPending}
-          className="w-full bg-blue-600 text-white py-2 rounded-lg text-sm font-semibold hover:bg-blue-700 disabled:opacity-50 transition-colors"
-        >
-          {tzMutation.isPending ? 'Saving…' : 'Save Timezone'}
-        </button>
-      </form>
-
-      {/* Login location card */}
-      <div className="bg-white rounded-lg shadow-sm border border-gray-200 p-6 space-y-4">
-        <h2 className="text-base font-semibold text-gray-700">Login Location</h2>
-        <p className="text-xs text-gray-500">
-          The automation logs into LinkedIn from our servers. Setting your location makes the
-          browser appear to come from where you normally log in, which reduces LinkedIn
-          "new location" security challenges.
-        </p>
-
-        <div className="text-sm text-gray-700">
-          <span className="font-medium text-gray-600">Current: </span>
-          {locationData?.latitude != null
-            ? `${[locationData.city, locationData.country].filter(Boolean).join(', ') || 'Set'} (${locationData.latitude.toFixed(3)}, ${locationData.longitude?.toFixed(3)})`
-            : 'Not set'}
-        </div>
-
-        {locationMsg && (
-          <p className={`text-sm font-medium ${locationMsg.ok ? 'text-green-600' : 'text-red-600'}`}>
-            {locationMsg.text}
-          </p>
-        )}
-
-        {/* Manual city/state entry */}
-        <div className="grid grid-cols-1 sm:grid-cols-3 gap-2">
-          <input
-            type="text"
-            value={locCity}
-            onChange={(e) => setLocCity(e.target.value)}
-            placeholder="City"
-            className="border border-gray-300 rounded-lg px-3 py-2 text-sm sm:col-span-1"
-          />
-          <input
-            type="text"
-            value={locState}
-            onChange={(e) => setLocState(e.target.value)}
-            placeholder="State / region"
-            className="border border-gray-300 rounded-lg px-3 py-2 text-sm sm:col-span-1"
-          />
-          <input
-            type="text"
-            value={locCountry}
-            onChange={(e) => setLocCountry(e.target.value)}
-            placeholder="Country (ISO-2)"
-            maxLength={2}
-            className="border border-gray-300 rounded-lg px-3 py-2 text-sm uppercase sm:col-span-1"
-          />
-        </div>
-        <button
-          type="button"
-          onClick={() => setLocationByCityMutation.mutate()}
-          disabled={setLocationByCityMutation.isPending || !locCity.trim()}
-          className="w-full bg-blue-600 text-white py-2 rounded-lg text-sm font-semibold hover:bg-blue-700 disabled:opacity-50 transition-colors"
-        >
-          {setLocationByCityMutation.isPending ? 'Setting…' : 'Set location'}
-        </button>
-
-        <div className="relative py-1">
-          <div className="border-t border-gray-200" />
-          <span className="absolute inset-0 -top-2 flex justify-center text-xs text-gray-400">
-            <span className="bg-white px-2">or</span>
-          </span>
-        </div>
-
-        <button
-          type="button"
-          onClick={() => autocaptureLocationMutation.mutate()}
-          disabled={autocaptureLocationMutation.isPending}
-          className="w-full bg-gray-100 text-gray-700 py-2 rounded-lg text-sm font-semibold hover:bg-gray-200 disabled:opacity-50 transition-colors"
-        >
-          {autocaptureLocationMutation.isPending ? 'Detecting…' : 'Use my current location (auto-detect)'}
-        </button>
+        ))}
       </div>
+
+      {activeTab === 'account' && (
+        <div className="space-y-6">
+          <SubscriptionCard />
+          <VideoCreditsCard />
+          <TimezoneCard />
+          <LoginLocationCard />
+          <InactivityCard />
+        </div>
+      )}
+
+      {activeTab === 'linkedin' && (
+        <div className="space-y-6">
+          <LinkedInLoginCard />
+          <CompanyPageCard />
+        </div>
+      )}
+
+      {activeTab === 'content' && (
+        <div className="space-y-6">
+          <ContentProfileCard />
+          <NewsletterCard />
+        </div>
+      )}
+
+      {activeTab === 'automation' && (
+        <div className="space-y-6">
+          <EngagementPreferencesCard />
+          <EngagementTargetingCard />
+          <GroupsCard />
+          <DmTemplatesCard />
+          <LeadMagnetCard />
+        </div>
+      )}
     </div>
   )
 }
